refactor(orchestrator): split LoopWarning out; add extraction prototype
- shared/types.ts passed the 500-LOC limit with the new skill fields, so
  LoopWarning moves to warning-types.ts and is re-exported (490 lines)
- static UX prototype of the five extraction states, linked from spec 18
