+++ plugins/sero-orchestrator-plugin/shared/types.ts
@@ -15,11 +15,11 @@ import type { LoopLibraryLink, StepOverride } from './library-types';
 import type { LogPolicy, UsageSummary } from './usage-types';
 import type { LoopWorkspaceRuntime, LoopWorkspaceSettings, ResolvedWorkspaceContext } from './workspace-types';
 import type { LoopInsight, LoopSuggestion } from './reflection-types';
-import type { LoopSkillLink, SkillDraft } from './skill-types';
 
-// Skill extraction types live in skill-types.ts (500-LOC limit); re-exported
-// here so existing imports from './types' keep resolving.
+// Skill extraction (specs/18) and loop warnings are split out the same way.
 export type { SkillDraft, SkillDraftStatus, LoopSkillLink } from './skill-types';
+import type { LoopSkillLink, SkillDraft } from './skill-types';
+import type { LoopWarning } from './warning-types';
 import type { AnsweredInput, HumanQuestion, PendingInput } from './human-input-types';
 import type { CompletionSignal, RecoveryDecision, RecoveryDecisionKind, StepCompletion } from './recovery-types';
 import type { Observation } from './observation-types';
@@ -177,22 +177,7 @@ export interface Loop {
   updatedAt: string;
 }
 
-export interface LoopWarning {
-  id: string;
-  code:
-    | 'mixed-workspace-targets'
-    | 'model-unavailable'
-    | 'agent-unavailable'
-    | 'event-chain-depth'
-    | 'event-dropped'
-    | 'event-queue-overflow'
-    | 'delivery-tool-missing'
-    | 'catalog-tool-missing';
-  message: string;
-  /** The step a runtime warning refers to (model/agent-unavailable), for de-duplication. */
-  stepId?: string;
-  createdAt: string;
-}
+export type { LoopWarning } from './warning-types';
 
 // ── Planning response & plan ────────────────────────────────
